stm32h7/linum-stm32h753bi: Add support to littlefs and nxffs with flash mem. via quadspi

+++ fs/nxffs/nxffs_dump.c
@@ -61,7 +61,7 @@ struct nxffs_blkinfo_s
 
 #if defined(CONFIG_DEBUG_FEATURES) && defined(CONFIG_DEBUG_FS)
 static const char g_hdrformat[] = "  BLOCK:OFFS  TYPE  STATE   LENGTH\n";
-static const char g_format[]    = "  %5d:%-5d %s %s %5d\n";
+static const char g_format[]    = "  %5"PRIi32":%-5d %s %s %5"PRIu32"\n";
 #endif
 
 /****************************************************************************
@@ -263,7 +263,8 @@ static inline ssize_t nxffs_analyzedata(FAR struct nxffs_blkinfo_s *blkinfo,
   if (crc != ecrc)
     {
       syslog(LOG_NOTICE, g_format,
-             blkinfo->block, offset, "DATA ", "CRC BAD", datlen);
+             blkinfo->block, offset, "DATA ", "CRC BAD",
+             (long unsigned int)datlen);
       return ERROR;
     }
 
@@ -272,7 +273,8 @@ static inline ssize_t nxffs_analyzedata(FAR struct nxffs_blkinfo_s *blkinfo,
   if (blkinfo->verbose)
     {
       syslog(LOG_NOTICE, g_format,
-             blkinfo->block, offset, "DATA ", "OK     ", datlen);
+             blkinfo->block, offset, "DATA ", "OK     ",
+             (long unsigned int)datlen);
     }
 
   return SIZEOF_NXFFS_DATA_HDR + datlen;
@@ -489,7 +491,7 @@ int nxffs_dump(FAR struct mtd_dev_s *mtd, bool verbose)
         }
     }
 
-  syslog(LOG_NOTICE, "%d blocks analyzed\n", blkinfo.nblocks);
+  syslog(LOG_NOTICE, "%" PRIi32 " blocks analyzed\n", blkinfo.nblocks);
 
   kmm_free(blkinfo.buffer);
   return OK;
